+++ crates/nu_plugin_polars/src/dataframe/command/datetime/mod.rs
@@ -15,7 +15,7 @@ mod get_year;
 mod replace_time_zone;
 mod strftime;
 mod truncate;
-mod date_range;
+mod date_ranges;
 
 use crate::PolarsPlugin;
 use nu_plugin::PluginCommand;
@@ -46,7 +46,7 @@ pub(crate) fn datetime_commands() -> Vec<Box<dyn PluginCommand<Plugin = PolarsPl
         Box::new(AsDate),
         Box::new(AsDateTime),
         Box::new(ConvertTimeZone),
-        Box::new(date_range::DateRange),
+        Box::new(date_ranges::DateRange),
         Box::new(ExprDatePart),
         Box::new(GetDay),
         Box::new(GetHour),
